test(codemod): fix Windows fixtures and classify native integration tests
