+++ scripts/test/test-semantic-audit-migration.ts
@@ -1523,18 +1523,6 @@ export const TEST_SEMANTIC_AUDIT_MIGRATION_ENTRIES:
         "tests/integration/semantic-unit-boundary/scripts/ci/windows-localhost.test.ts",
       "removalPr": "PR 4f",
     }),
-    entry("scripts/codemods/migrate-esm-sh-imports.test.ts", [
-      "filesystem-read",
-      "filesystem-write",
-    ], {
-      "disposition": "integration-relocation",
-      "owner": "scripts-tooling",
-      "rationale":
-        "Exercises filesystem mutation, process, server, network, browser, or multi-component runtime behavior outside the colocated unit boundary.",
-      "destination":
-        "tests/integration/semantic-unit-boundary/scripts/codemods/migrate-esm-sh-imports.test.ts",
-      "removalPr": "PR 4f",
-    }),
     entry("scripts/docs/generate-api-reference.test.ts", [
       "filesystem-read",
       "filesystem-write",
